+++ packages/discovery/src/index.ts
@@ -1,4 +1,5 @@
 export * from "./catalog";
+export * from "./anm";
 export * from "./camuv";
 export * from "./cdnod";
 export * from "./contracts";